// src/app/model/delivery.ts
export class Delivery {
	number: string;
    name: string;
    documentDate: string;
    documentDateStamp: number;
    deliveryFile: string;
    paymentFile: string;
    accessories: number;
    books: number;
    cars: number;
    coaches: number;
    elements: number;
    id: number;
    locs: number;
    toPrint: number;
    type: number;
}